refactor: Restructure frontend service namespaces
