+++ Spreeview/SpreeviewFrontend/SpreeviewFrontend/Components/Pages/Logout.razor
@@ -1,6 +1,6 @@
 ﻿@page "/logout"
 @using SpreeviewAPI.Wrappers
-@using SpreeviewFrontend.Client.Identity
+@using SpreeviewFrontend.Services.AccountManagement
 @inject IAccountManagementService AccountManagementService
 
 <PageTitle>Logout</PageTitle>
