+++ Spreeview/SpreeviewFrontend/SpreeviewFrontend/Components/Pages/Register.razor
@@ -1,7 +1,7 @@
 ﻿@page "/register"
 @using RecordShop.Frontend.Client.Identity.Models
 @using SpreeviewAPI.Wrappers
-@using SpreeviewFrontend.Client.Identity
+@using SpreeviewFrontend.Services.AccountManagement
 @using SpreeviewFrontend.Components.Authentication
 @inject IAccountManagementService AccountManagementService
