+++ Spreeview/SpreeviewFrontend/SpreeviewFrontend/Services/ApiHealth/ApiHealthService.cs
@@ -3,7 +3,7 @@ using CommonLibrary.DataClasses.ApiHealthModel;
 using CommonLibrary.DataClasses.ApiHealthModel;
 using SpreeviewAPI.Wrappers;
 
-namespace SpreeviewFrontend.Services.HealthCheck;
+namespace SpreeviewFrontend.Services.ApiHealth;
 
 public class ApiHealthService : IApiHealthService
 {
@@ -20,7 +20,7 @@ public class ApiHealthService : IApiHealthService
         _logger = logger;
     }
 
-    public async Task<ServiceObjectResponse<ApiHealth>> GetHealthAsync()
+    public async Task<ServiceObjectResponse<CommonLibrary.DataClasses.ApiHealthModel.ApiHealth>> GetHealthAsync()
     {
         try
         {
@@ -30,7 +30,7 @@ public class ApiHealthService : IApiHealthService
             // If not successful
             if (!result.IsSuccessStatusCode)
             {
-                return new ServiceObjectResponse<ApiHealth>()
+                return new ServiceObjectResponse<CommonLibrary.DataClasses.ApiHealthModel.ApiHealth>()
                 {
                     Type = ServiceResponseType.Failure,
                     Messages = ["Failed to get server health. Endpoint response did not indicate success."]
@@ -41,15 +41,15 @@ public class ApiHealthService : IApiHealthService
             var content = await result.Content.ReadAsStringAsync();
         
             // Deserialize the result
-            var health = JsonSerializer.Deserialize<ApiHealth>(content, _jsonSerializerOptions);
+            var health = JsonSerializer.Deserialize<CommonLibrary.DataClasses.ApiHealthModel.ApiHealth>(content, _jsonSerializerOptions);
         
             // Success
-            return new ServiceObjectResponse<ApiHealth>() { Type = ServiceResponseType.Success , Value = health};
+            return new ServiceObjectResponse<CommonLibrary.DataClasses.ApiHealthModel.ApiHealth>() { Type = ServiceResponseType.Success , Value = health};
         }
         catch (Exception e)
         {
             _logger.LogInformation("Could not check health due to server error.");
-            return new ServiceObjectResponse<ApiHealth>();
+            return new ServiceObjectResponse<CommonLibrary.DataClasses.ApiHealthModel.ApiHealth>();
         }
     }
 }
